refactor: remove IR emitter implementation and related functionality
- Deleted `ir_emitter`, `ir_emitter_memento`, and associated header/source files.
- Removed specific visit methods and context expansion logic.
- Discontinued runtime emission through `c4_runtime_emitter`.

+++ src/c4/src/p2/lex/match_helper.hxx
@@ -43,6 +43,7 @@
 #include <string_view>
 #include <uni_algo/ranges_conv.h>
 #include <iterator>
+#include <libassert/assert.hpp>
 
 namespace c4 {
     struct position;
@@ -51,7 +52,7 @@ namespace c4 {
 inline std::size_t
 utf8_strlen(std::string_view str) {
     auto str_view = str | una::views::utf8;
-    return std::distance(str_view.begin(), str_view.end());
+    return static_cast<std::size_t>(std::distance(str_view.begin(), str_view.end()));
 }
 
 struct line_data {
